… dbms/tests/queries/0_stateless/00517_date_parsing.reference
… dbms/tests/queries/0_stateless/00517_date_parsing.sql