Accepting datetime string from 2000-1-1 1:2:3 to 2000-01-01 01:02:03.