@alexey-milovidov 2017-12-27
I wonder, what is the main motivation for this change?