@amosbird 2017-12-27
The main purpose is to accept flexible datetime formats of data which I just received :-)