@alexey-milovidov 2017-12-28
What about performance tests?

If the performance decrease (at least 1%), it's important to choose, for what feature we will trade the performance.

For example, I am ready to trade few percent of performance to add support of parsing DateTime in full ISO 8601 format like `2017-12-28T15:01:29+00:00` and `2017-12-28T15:01:29Z`. This seems as much more relevant feature.

As next step, I consider adding switchable DateTime parsing. For example, what if you can choose (with settings) between simple and fast parsing and customizeable parsing with user-defined format? (And that user-defined format can include your case.)