@amosbird 2017-12-28
## Before this patch
```
QPS: 2.528, RPS: 257865127.011, MiB/s: 983.677, result RPS: 2.528, result MiB/s: 0.000.
0.000%  0.363 sec.
10.000% 0.369 sec.
20.000% 0.370 sec.
30.000% 0.372 sec.
40.000% 0.374 sec.
50.000% 0.377 sec.
60.000% 0.398 sec.
70.000% 0.414 sec.
80.000% 0.424 sec.
90.000% 0.441 sec.
95.000% 0.451 sec.
99.000% 0.467 sec.
99.900% 0.474 sec.
99.990% 0.475 sec.
```

## After this patch
```
Queries executed: 204.

QPS: 2.486, RPS: 253612596.672, MiB/s: 967.455, result RPS: 2.486, result MiB/s: 0.000.
0.000%  0.361 sec.
10.000% 0.370 sec.
20.000% 0.373 sec.
30.000% 0.375 sec.
40.000% 0.378 sec.
50.000% 0.402 sec.
60.000% 0.411 sec.
70.000% 0.424 sec.
80.000% 0.431 sec.
90.000% 0.444 sec.
95.000% 0.452 sec.
99.000% 0.457 sec.
99.900% 0.463 sec.
99.990% 0.464 sec.
```

It doesn't seem to have any performance impact.

> As next step, I consider adding switchable DateTime parsing. 

That sounds great!