@alexey-milovidov 2018-02-13
First step: #1884.
(Currently it doesn't change the default behaviour, only introduces a new function)