@krzyzakp 2018-06-05
@alexey-milovidov any chance to use function, which you introduced here? Would be super useful for us.