@alexey-milovidov 2018-06-05
I want to add a setting to switch between basic and "best effort" parsing in text formats.
This looks like relatively easy task but it's not in high priority right now.