@alexey-milovidov 2018-06-08
Now you can use setting
`SET date_time_input_format = 'best_effort'`
to enable extended DateTime parsing.

Notes:
- it doesn't affect the behaviour of type conversion functions;